The query method between Account and EducationProgram has been changed. (#7)

+++ Business/Dtos/Requests/AccountEducationProgramRequests/UpdateAccountEducationProgramRequest.cs
@@ -6,5 +6,5 @@ public class UpdateAccountEducationProgramRequest
     public Guid AccountId { get; set; }
     public Guid EducationProgramId { get; set; }
     public double StatusPercent { get; set; }
-    public double TimeSpent { get; set; }
+    public int TimeSpent { get; set; }
 }
