+++ Business/Dtos/Responses/AccountEducationProgramResponses/DeletedAccountEducationProgramResponse.cs
@@ -6,5 +6,5 @@ public class DeletedAccountEducationProgramResponse
     public Guid AccountId { get; set; }
     public Guid EducationProgramId { get; set; }
     public double StatusPercent { get; set; }
-    public double TimeSpent { get; set; }
+    public int TimeSpent { get; set; }
 }
